+++ db/schema.rb
@@ -10,7 +10,7 @@
 #
 # It's strongly recommended that you check this file into your version control system.
 
-ActiveRecord::Schema.define(version: 2020_03_14_101911) do
+ActiveRecord::Schema.define(version: 2020_03_26_072425) do
 
   # These are extensions that must be enabled in order to support this database
   enable_extension "pg_stat_statements"
@@ -122,6 +122,7 @@ ActiveRecord::Schema.define(version: 2020_03_14_101911) do
     t.integer "item_level", default: 0, null: false
     t.boolean "item_level_calculated", default: false, null: false
     t.string "current_roles", default: [], array: true
+    t.boolean "main", default: false
     t.index ["character_class_id"], name: "index_characters_on_character_class_id"
     t.index ["guild_id"], name: "index_characters_on_guild_id"
     t.index ["name"], name: "index_characters_on_name"